requirements.txt
python-telegram-bot==13.15
pyyaml
deep-translator==1.9.2
llama-cpp-python